Move RpcPubSubService into its own module
